Potential fix for #BA1660 + ZoneType refactoring
Synchronized access to counter variables to ensure `onLeaveFlyArea()` will be called when gliding out of a fly zone. If the client thinks you're in fly mode outside a fly zone, it kicks players in the top left corner of the map when touching the ground (see https://beyond-aion.com/forum/thread/1660-gleitbug/).
Removed unused ZoneTypes and refactored `//zone` command.