Removed unnecessary effect restrictions and observer types
This fixes Cursecloud and Curse of Weakness did not inflict bonus damage to npcs casting magic spells.
The removal of `ObserverType.NPCDIALOGREQUEST` will not be noticeable, as it was only used in HideEffect and currently all npc_template.xml entries have can_talk_invisible="false", which is being checked since d2fd95d.